Fixed auth for UpdateLinkBundle endpoint
The most significant changes in the code are related to the interaction with the Cosmos DB container and the introduction of additional security checks. The code now retrieves the `ClientPrincipal` from the request and performs a query on the Cosmos DB container to find a matching `LinkBundle`. If a match is found, the code checks if the hashed username and the `IdentityProvider` match the `UserId` and `Provider` of the `LinkBundle`. If they do not match, an `Unauthorized` HTTP response is returned. If they do match, the `UserId` and `Provider` of the `LinkBundle` are updated and the `LinkBundle` is upserted into the Cosmos DB container. If no match is found, a `NotFound` HTTP response is returned.

Changes:

1. The `using` statements at the top of the file have been expanded to include additional namespaces such as `Api.Utility`, `BlazorApp.Shared`, `Microsoft.Azure.Cosmos`, `Microsoft.Azure.Functions.Worker`, `Microsoft.Azure.Functions.Worker.Http`, `Microsoft.Extensions.Logging`, `System`, and `System.Linq`.
2. The constructor for the `UpdateLinkBundle` class has been modified to change the parameter name from `_cosmosClient` to `cosmosClient`.
3. The code for interacting with the Cosmos DB container has been updated. It now retrieves the `ClientPrincipal` from the request and performs a query on the Cosmos DB container to find a matching `LinkBundle`.
4. If a matching `LinkBundle` is found, the code checks if the hashed username and the `IdentityProvider` match the `UserId` and `Provider` of the `LinkBundle`. If they do not match, an `Unauthorized` HTTP response is returned.
5. If the `UserId` and `Provider` do match, the `UserId` and `Provider` of the `LinkBundle` are updated and the `LinkBundle` is upserted into the Cosmos DB container.
6. If no matching `LinkBundle` is found in the Cosmos DB container, then an HTTP response with a status code of `NotFound` is returned.