Improve title assignment in GetOpenGraphInfo.cs
The commit addresses an issue in the `GetOpenGraphInfo.cs` file in the `Api.Functions` namespace. Previously, the `title` variable was mistakenly assigned the content of the `twitter:title` meta tag twice in a row if the previous selectors didn't find any content. The updated code now checks the text content of the `h1` tag before checking the `og:site_name` meta tag. This change will likely provide a more accurate `title` when the previous selectors don't find any content.