Remove EditForm and Regular Expressions to cut dependencies.
Refactor URL validation and form handling in LinkBundle and NewLink

Refactored the validation process for the `VanityUrl` property in the `LinkBundle` class to check for letters, numbers, and dashes without using regex. The `NewLink.razor` file was significantly refactored, replacing the `EditForm` component with a standard HTML form and the `OnValidSubmit` method with an `OnSubmit` method that includes a URL validation check. The `Link` and `LinkBundle` classes no longer use data annotations or regular expressions for URL validation. Minor refactoring was also done on the `GetOpenGraphInfoForLink` method in `NewLink.razor`, mainly for code formatting and indentation.